throws implemented in service implementation

+++ src/main/java/com/alkemy/challenge/service/impl/PostServiceImpl.java
@@ -2,16 +2,25 @@ package com.alkemy.challenge.service.impl;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import javax.persistence.EntityNotFoundException;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import com.alkemy.challenge.Exception.BodyDtoException;
+import com.alkemy.challenge.Exception.msg.ErrorMsg;
 import com.alkemy.challenge.dto.PostDto;
 import com.alkemy.challenge.dto.PostDtoGetAll;
 import com.alkemy.challenge.mapper.PostMapper;
 import com.alkemy.challenge.model.Post;
 import com.alkemy.challenge.model.Specification.PostSpecification;
 import com.alkemy.challenge.repository.PostRepository;
+import com.alkemy.challenge.security.model.User;
+import com.alkemy.challenge.security.service.UserService;
 import com.alkemy.challenge.service.PostService;
+import com.alkemy.challenge.validator.DtoValidator;
 
+@Service
 public class PostServiceImpl implements PostService {
 
 	@Autowired
@@ -23,14 +32,31 @@ public class PostServiceImpl implements PostService {
 	@Autowired
 	PostSpecification postSpecification;
 
+	@Autowired
+	DtoValidator dtoValidator;
+
+	@Autowired
+	UserService userService;
+
 	@Override
-	public PostDto savePost(PostDto postDto) {
+	public PostDto savePost(PostDto postDto, String username) {
 
-		Post post = postMapper.toPost(postDto);
+		// it wont be null because the security filter previously took care of validate
+		// the jwt
+		User user = userService.findUserByEmail(username);
 
-		post = postRepository.save(post);
+		if (dtoValidator.postDtoIsValid(postDto)) {
 
-		return postMapper.toPostDto(post);
+			Post post = postMapper.toPost(postDto, user.getUserId());
+
+			post = postRepository.save(post);
+
+			return postMapper.toPostDto(post);
+
+		}
+
+		throw new BodyDtoException(ErrorMsg.dtoBodyNotValidMsg());
+
 	}
 
 	@Override
@@ -52,19 +78,36 @@ public class PostServiceImpl implements PostService {
 	@Override
 	public PostDto updatePost(PostDto postDto, Long id) {
 
-		Post post = postRepository.getById(id);
+		if (dtoValidator.postDtoIsValid(postDto)) {
 
-		postMapper.refreshValues(post, postDto);
+			Post post = postRepository.getById(id);
 
-		post = postRepository.save(post);
+			postMapper.refreshValues(post, postDto);
 
-		return postMapper.toPostDto(post);
+			post = postRepository.save(post);
+
+			return postMapper.toPostDto(post);
+
+		} else {
+
+			throw new BodyDtoException(ErrorMsg.dtoBodyNotValidMsg());
+
+		}
+
 	}
 
 	@Override
 	public void deletePost(Long id) {
 
-		postRepository.deleteById(id);
+		if (postRepository.existsById(id)) {
+
+			postRepository.deleteById(id);
+
+		} else {
+
+			throw new EntityNotFoundException(ErrorMsg.postNotFoundMsg(id));
+
+		}
 
 	}
